requirements.txt
Django==3.0.4
Pillow==7.0.0
mysqlclient==1.4.6
django-crispy-forms==1.9.0
# Django libsass does not support Django 3.0 yet
#git+git://github.com/gafderks/django-libsass@master#egg=django-libsass
django_compressor==2.4
django-money==1.0
django-admin-sortable==2.2.3
